+++ hooks/post_apply.py
@@ -95,7 +95,11 @@ def main(
 
 if __name__ == "__main__":
     app_interface_input = parse_model(AppInterfaceInput, read_input_from_file())
-    plan = TerraformJsonPlanParser(plan_path=sys.argv[1])
+    plan_json = os.environ.get("PLAN_FILE_JSON")
+    if not plan_json:
+        logger.error("PLAN_FILE_JSON environment variable not set")
+        sys.exit(1)
+    plan = TerraformJsonPlanParser(plan_path=plan_json)
     main(
         plan,
         app_interface_input,
