+++ tests/hooks/test_hooks_post_output.py
@@ -1,6 +1,6 @@
 import pytest
 
-from hooks.check_output import check
+from hooks.post_output import check
 
 
 @pytest.mark.parametrize(
